cookbook/integrations/flytekit_plugins/sql/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    /Library/Developer/CommandLineTools/usr/bin/make requirements.txt
#
attrs==21.2.0
    # via scantree
certifi==2021.5.30
    # via requests
charset-normalizer==2.0.1
    # via requests
click==7.1.2
    # via flytekit
croniter==1.0.15
    # via flytekit
cycler==0.10.0
    # via matplotlib
dataclasses-json==0.5.4
    # via flytekit
decorator==5.0.9
    # via retry
deprecated==1.2.12
    # via flytekit
dirhash==0.2.1
    # via flytekit
docker-image-py==0.1.10
    # via flytekit
flyteidl==0.19.13
    # via flytekit
flytekit==0.20.1
    # via
    #   -r ../../../common/requirements-common.in
    #   flytekitplugins-sqlalchemy
flytekitplugins-sqlalchemy==0.20.1
    # via -r requirements.in
greenlet==1.1.0
    # via sqlalchemy
grpcio==1.38.1
    # via flytekit
idna==3.2
    # via requests
importlib-metadata==4.6.1
    # via keyring
keyring==23.0.1
    # via flytekit
kiwisolver==1.3.1
    # via matplotlib
marshmallow==3.12.2
    # via
    #   dataclasses-json
    #   marshmallow-enum
    #   marshmallow-jsonschema
marshmallow-enum==1.5.1
    # via dataclasses-json
marshmallow-jsonschema==0.12.0
    # via flytekit
matplotlib==3.4.2
    # via -r ../../../common/requirements-common.in
mypy-extensions==0.4.3
    # via typing-inspect
natsort==7.1.1
    # via flytekit
numpy==1.21.0
    # via
    #   matplotlib
    #   pandas
    #   pyarrow
pandas==1.3.0
    # via flytekit
pathspec==0.8.1
    # via scantree
pillow==8.3.1
    # via matplotlib
protobuf==3.17.3
    # via
    #   flyteidl
    #   flytekit
py==1.10.0
    # via retry
pyarrow==3.0.0
    # via flytekit
pyparsing==2.4.7
    # via matplotlib
python-dateutil==2.8.1
    # via
    #   croniter
    #   flytekit
    #   matplotlib
    #   pandas
python-json-logger==2.0.1
    # via flytekit
pytimeparse==1.1.8
    # via flytekit
pytz==2018.4
    # via
    #   flytekit
    #   pandas
regex==2021.7.6
    # via docker-image-py
requests==2.26.0
    # via
    #   flytekit
    #   responses
responses==0.13.3
    # via flytekit
retry==0.9.2
    # via flytekit
scantree==0.0.1
    # via dirhash
six==1.16.0
    # via
    #   cycler
    #   flytekit
    #   grpcio
    #   protobuf
    #   python-dateutil
    #   responses
    #   scantree
sortedcontainers==2.4.0
    # via flytekit
sqlalchemy==1.4.20
    # via flytekitplugins-sqlalchemy
statsd==3.3.0
    # via flytekit
stringcase==1.2.0
    # via dataclasses-json
typing-extensions==3.10.0.0
    # via typing-inspect
typing-inspect==0.7.1
    # via dataclasses-json
urllib3==1.26.6
    # via
    #   flytekit
    #   requests
    #   responses
wheel==0.36.2
    # via
    #   -r ../../../common/requirements-common.in
    #   flytekit
wrapt==1.12.1
    # via
    #   deprecated
    #   flytekit
zipp==3.5.0
    # via importlib-metadata
